# lib/api-zod/src/generated/types/ticketAssign.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Imperium website API
 * OpenAPI spec version: 0.1.0
 */

export interface TicketAssign {
  /** @nullable */
  staffId: number | null;
}